Make GenomeBrowsers render in jupyter notebook
This commit updates all the code to the point a browser will render
in a jupyter notebook.

Other updates and tweaks are still required to get it working smoothly.
