+++ REQUIRE
@@ -1,3 +1,5 @@
 julia 0.6
 HttpCommon
 HttpServer
+JSON
+GenomicFeatures
