+++ src/browser.jl
@@ -1,0 +1,5 @@
+
+struct DallianceBrowser
+    genome_config::Dict
+    sources::Vector{Dict}
+end
